[REFACTOR] Reorganize Interfaces by Domain
