+++ interface/calculator.ts
@@ -23,4 +23,36 @@ export interface BadmintonCostCalculatorProps {
   onClose: () => void;
   players: PlayerStats[];
   onBackToResults: () => void;
+}
+
+export interface ShuttlecockSectionProps {
+  shuttlecock: Shuttlecock;
+  onShuttlecockChange: (shuttlecock: Shuttlecock) => void;
+}
+
+export interface CostBreakdownProps {
+  players: PlayerStats[];
+  courtFee: CourtFee;
+  shuttlecock: Shuttlecock;
+  customExpenses: CustomExpense[];
+  totalCost: number;
+}
+
+export interface AdditionalExpensesProps {
+  customExpenses: CustomExpense[];
+  players: PlayerStats[];
+  onCustomExpensesChange: (expenses: CustomExpense[]) => void;
+}
+
+export interface PlayerCost {
+  name: string;
+  sharedCost: number;
+  customExpenses: number;
+  total: number;
+  hours?: number;
+}
+
+export interface CourtFeeSectionProps {
+  courtFee: CourtFee;
+  onCourtFeeChange: (courtFee: CourtFee) => void;
 }
