+++ interface/index.ts
@@ -1,3 +1,5 @@
 export * from './player';
 export * from './match';
-export * from './calculator'; 
+export * from './calculator';
+export * from './qrcode';
+export * from './modal';
