+++ interface/match.ts
@@ -11,4 +11,14 @@ export interface Match {
 export interface MatchHistory {
   matches: Match[];
   players: Player[];
+}
+
+export interface CurrentMatchProps {
+  match: Match;
+  matchNumber: number;
+  selectedWinner: 'team1' | 'team2' | null;
+  onSelectWinner: (team: 'team1' | 'team2') => void;
+  onConfirmWinner: (matchId: string) => void;
+  onRevertMatch?: (matchId: string) => void;
+  showRevert?: boolean;
 }
